refactor: split transform manager utils
