+++ src/fold_db_core/transform_manager/utils/mod.rs
@@ -15,12 +15,12 @@ use crate::fold_db_core::infrastructure::message_bus::{
     MessageBus,
     schema_events::TransformExecuted,
 };
-use crate::schema::types::{SchemaError, Schema, Transform};
+use crate::schema::types::{SchemaError, Schema};
 use crate::schema::types::field::variant::FieldVariant;
 use crate::schema::types::field::common::Field;
 use serde_json::Value as JsonValue;
 use log::{info, error, warn};
-use std::sync::{Arc, RwLock, RwLockReadGuard, RwLockWriteGuard};
+use std::sync::Arc;
 use std::collections::{HashMap, HashSet};
 
 // Re-export commonly used types to avoid import duplication
@@ -29,333 +29,19 @@ pub use serde_json::Value;
 /// Single unified utility combining ALL transform manager utilities
 pub struct TransformUtils;
 
+pub mod conversion;
+pub mod default_values;
+pub mod locking;
+pub mod serialization;
+pub mod validation;
+
+pub use conversion::*;
+pub use default_values::*;
+pub use locking::*;
+pub use serialization::*;
+pub use validation::*;
+
 impl TransformUtils {
-    // ========== CONVERSION UTILITIES ==========
-    
-    /// Convert JsonValue to Value with unified error handling
-    pub fn json_to_value(json_value: JsonValue) -> Result<JsonValue, SchemaError> {
-        Ok(json_value)
-    }
-
-    /// Validate and convert JsonValue with type checking
-    pub fn validate_and_convert(
-        json_value: JsonValue,
-        expected_type: &str,
-        field_name: &str,
-    ) -> Result<JsonValue, SchemaError> {
-        info!("🔄 Converting field '{}' (expected type: {})", field_name, expected_type);
-        
-        let is_valid = match expected_type.to_lowercase().as_str() {
-            "string" | "str" => json_value.is_string(),
-            "number" | "integer" | "int" => json_value.is_number() && json_value.as_i64().is_some(),
-            "float" | "double" => json_value.is_number() && json_value.as_f64().is_some(),
-            "boolean" | "bool" => json_value.is_boolean(),
-            "array" => json_value.is_array(),
-            "object" => json_value.is_object(),
-            "null" => json_value.is_null(),
-            _ => {
-                warn!("⚠️ Unknown expected type '{}' for field '{}', allowing any type", expected_type, field_name);
-                true
-            }
-        };
-
-        if !is_valid {
-            let error_msg = format!(
-                "Type validation failed for field '{}': expected '{}', got '{:?}'",
-                field_name, expected_type, json_value
-            );
-            error!("❌ {}", error_msg);
-            return Err(SchemaError::InvalidData(error_msg));
-        }
-
-        info!("✅ Successfully validated and converted field '{}'", field_name);
-        Ok(json_value)
-    }
-
-    /// Convert string to JsonValue with type inference
-    pub fn string_to_json_value(value_str: &str, infer_type: bool) -> JsonValue {
-        if !infer_type {
-            return JsonValue::String(value_str.to_string());
-        }
-
-        // Try to infer the type from the string content
-        if value_str.eq_ignore_ascii_case("true") {
-            return JsonValue::Bool(true);
-        }
-        
-        if value_str.eq_ignore_ascii_case("false") {
-            return JsonValue::Bool(false);
-        }
-        
-        if value_str.eq_ignore_ascii_case("null") {
-            return JsonValue::Null;
-        }
-
-        // Try to parse as integer
-        if let Ok(int_val) = value_str.parse::<i64>() {
-            return JsonValue::Number(serde_json::Number::from(int_val));
-        }
-
-        // Try to parse as float
-        if let Ok(float_val) = value_str.parse::<f64>() {
-            if let Some(num) = serde_json::Number::from_f64(float_val) {
-                return JsonValue::Number(num);
-            }
-        }
-
-        JsonValue::String(value_str.to_string())
-    }
-
-    // ========== DEFAULT VALUE UTILITIES ==========
-    
-    /// Enhanced default value generation with comprehensive field mapping
-    pub fn get_default_value_for_field(field_name: &str) -> JsonValue {
-        info!("📊 Generating default value for field: {}", field_name);
-        
-        let default_value = match field_name {
-            // Common input field names
-            "input1" => JsonValue::Number(serde_json::Number::from(42)),
-            "input2" => JsonValue::Number(serde_json::Number::from(24)),
-            "value1" => JsonValue::Number(serde_json::Number::from(5)),
-            "value2" => JsonValue::Number(serde_json::Number::from(10)),
-            
-            // Common calculation inputs
-            "weight" => JsonValue::Number(serde_json::Number::from(70)),
-            "height" => JsonValue::Number(serde_json::Number::from_f64(1.75).unwrap_or(serde_json::Number::from(175))),
-            "age" => JsonValue::Number(serde_json::Number::from(30)),
-            
-            // Common identifiers
-            "id" | "user_id" | "patient_id" => JsonValue::String("default_id".to_string()),
-            "name" | "username" | "patient_name" => JsonValue::String("default_name".to_string()),
-            
-            // Common boolean flags
-            "active" | "enabled" | "is_valid" => JsonValue::Bool(true),
-            "disabled" | "inactive" | "is_deleted" => JsonValue::Bool(false),
-            
-            // Common numeric defaults
-            "score" | "rating" => JsonValue::Number(serde_json::Number::from(0)),
-            "count" | "quantity" => JsonValue::Number(serde_json::Number::from(1)),
-            "price" | "amount" => JsonValue::Number(serde_json::Number::from_f64(0.0).unwrap()),
-            
-            // Fallback for unknown fields based on heuristics
-            _ => Self::infer_default_by_name_pattern(field_name)
-        };
-        
-        info!("📊 Default value for '{}': {}", field_name, default_value);
-        default_value
-    }
-
-    /// Infer default value based on field name patterns
-    fn infer_default_by_name_pattern(field_name: &str) -> JsonValue {
-        let lower_name = field_name.to_lowercase();
-        
-        match true {
-            _ if lower_name.contains("count") || lower_name.contains("number") || lower_name.contains("value") => 
-                JsonValue::Number(serde_json::Number::from(0)),
-            _ if lower_name.contains("active") || lower_name.contains("enabled") || lower_name.contains("valid") => 
-                JsonValue::Bool(false),
-            _ if lower_name.contains("list") || lower_name.contains("array") || lower_name.contains("tags") => 
-                JsonValue::Array(vec![]),
-            _ if lower_name.contains("config") || lower_name.contains("meta") || lower_name.contains("data") => 
-                JsonValue::Object(serde_json::Map::new()),
-            _ => JsonValue::String("default".to_string()),
-        }
-    }
-
-    /// Infer type from field name for smart defaults
-    pub fn infer_type_from_field_name(field_name: &str) -> &'static str {
-        let lower_name = field_name.to_lowercase();
-        match true {
-            _ if lower_name.contains("id") || lower_name.contains("name") || lower_name.contains("email") => "string",
-            _ if lower_name.contains("count") || lower_name.contains("age") || lower_name.contains("score") => "integer",
-            _ if lower_name.contains("weight") || lower_name.contains("height") || lower_name.contains("price") => "float",
-            _ if lower_name.contains("active") || lower_name.contains("enabled") || lower_name.contains("valid") => "boolean",
-            _ if lower_name.contains("list") || lower_name.contains("items") || lower_name.contains("tags") => "array",
-            _ if lower_name.contains("config") || lower_name.contains("settings") || lower_name.contains("meta") => "object",
-            _ => "string" // Default to string
-        }
-    }
-
-    /// Get typed default value based on expected type
-    pub fn get_typed_default_value(field_name: &str, expected_type: &str) -> JsonValue {
-        info!("📊 Generating typed default value for field: {} (type: {})", field_name, expected_type);
-        
-        let default_value = match expected_type.to_lowercase().as_str() {
-            "string" | "str" => JsonValue::String(format!("default_{}", field_name)),
-            "number" | "integer" | "int" => JsonValue::Number(serde_json::Number::from(0)),
-            "float" | "double" => JsonValue::Number(serde_json::Number::from_f64(0.0).unwrap()),
-            "boolean" | "bool" => JsonValue::Bool(false),
-            "array" => JsonValue::Array(vec![]),
-            "object" => JsonValue::Object(serde_json::Map::new()),
-            "null" => JsonValue::Null,
-            _ => {
-                warn!("⚠️ Unknown type '{}' for field '{}', using field-based default", expected_type, field_name);
-                Self::get_default_value_for_field(field_name)
-            }
-        };
-        
-        info!("📊 Typed default value for '{}' ({}): {}", field_name, expected_type, default_value);
-        default_value
-    }
-
-    // ========== VALIDATION UTILITIES ==========
-    
-    /// Comprehensive transform validation
-    pub fn validate_transform_registration(
-        transform_id: &str,
-        transform: &Transform,
-    ) -> Result<(), SchemaError> {
-        info!("🔍 Validating transform registration for: {}", transform_id);
-
-        // Validate transform ID is not empty
-        if transform_id.trim().is_empty() {
-            let error_msg = "Transform ID cannot be empty".to_string();
-            error!("❌ {}", error_msg);
-            return Err(SchemaError::InvalidData(error_msg));
-        }
-
-        // Validate transform has inputs
-        let inputs = transform.get_inputs();
-        if inputs.is_empty() {
-            let error_msg = format!("Transform '{}' must have at least one input", transform_id);
-            error!("❌ {}", error_msg);
-            return Err(SchemaError::InvalidData(error_msg));
-        }
-
-        // Validate transform has output
-        let output = transform.get_output();
-        if output.trim().is_empty() {
-            let error_msg = format!("Transform '{}' must have a valid output field", transform_id);
-            error!("❌ {}", error_msg);
-            return Err(SchemaError::InvalidData(error_msg));
-        }
-
-        // Validate transform logic is not empty
-        if transform.logic.trim().is_empty() {
-            let error_msg = format!("Transform '{}' must have non-empty logic", transform_id);
-            error!("❌ {}", error_msg);
-            return Err(SchemaError::InvalidData(error_msg));
-        }
-
-        info!("✅ Transform registration validation passed for: {}", transform_id);
-        Ok(())
-    }
-
-    /// Validate field name format
-    pub fn validate_field_name(field_name: &str, context: &str) -> Result<(), SchemaError> {
-        info!("🔍 Validating field name '{}' in context: {}", field_name, context);
-
-        if field_name.trim().is_empty() {
-            let error_msg = format!("Field name cannot be empty in context: {}", context);
-            error!("❌ {}", error_msg);
-            return Err(SchemaError::InvalidData(error_msg));
-        }
-
-        // Check for valid schema.field format
-        let parts: Vec<&str> = field_name.split('.').collect();
-        if parts.len() != 2 {
-            let error_msg = format!(
-                "Field name '{}' must be in format 'schema.field' in context: {}",
-                field_name, context
-            );
-            error!("❌ {}", error_msg);
-            return Err(SchemaError::InvalidData(error_msg));
-        }
-
-        let (schema_name, field_name_part) = (parts[0], parts[1]);
-        
-        if schema_name.trim().is_empty() || field_name_part.trim().is_empty() {
-            let error_msg = format!("Schema and field names cannot be empty in field '{}' (context: {})", field_name, context);
-            error!("❌ {}", error_msg);
-            return Err(SchemaError::InvalidData(error_msg));
-        }
-
-        info!("✅ Field name validation passed for: {}", field_name);
-        Ok(())
-    }
-
-    // ========== LOCK UTILITIES ==========
-    
-    /// Acquire a read lock with consistent error handling
-    pub fn read_lock<'a, T>(lock: &'a RwLock<T>, lock_name: &str) -> Result<RwLockReadGuard<'a, T>, SchemaError> {
-        lock.read().map_err(|_| {
-            SchemaError::InvalidData(format!("Failed to acquire {} read lock", lock_name))
-        })
-    }
-    
-    /// Acquire a write lock with consistent error handling
-    pub fn write_lock<'a, T>(lock: &'a RwLock<T>, lock_name: &str) -> Result<RwLockWriteGuard<'a, T>, SchemaError> {
-        lock.write().map_err(|_| {
-            SchemaError::InvalidData(format!("Failed to acquire {} write lock", lock_name))
-        })
-    }
-
-    // ========== SERIALIZATION UTILITIES ==========
-    
-    /// Serialize a mapping to bytes with consistent error handling
-    pub fn serialize_mapping<T>(
-        mapping: &RwLock<T>,
-        mapping_name: &str,
-    ) -> Result<Vec<u8>, SchemaError>
-    where
-        T: serde::Serialize,
-    {
-        info!("🔄 Serializing mapping: {}", mapping_name);
-        
-        let map = Self::read_lock(mapping, mapping_name)?;
-        let json = serde_json::to_vec(&*map).map_err(|e| {
-            let error_msg = format!("Failed to serialize {}: {}", mapping_name, e);
-            error!("❌ {}", error_msg);
-            SchemaError::InvalidData(error_msg)
-        })?;
-        
-        info!("✅ Successfully serialized mapping: {} ({} bytes)", mapping_name, json.len());
-        Ok(json)
-    }
-
-    /// Deserialize mapping data with consistent error handling
-    pub fn deserialize_mapping<T>(
-        data: &[u8],
-        mapping_name: &str,
-    ) -> Result<T, SchemaError>
-    where
-        T: serde::de::DeserializeOwned + Default,
-    {
-        info!("🔄 Deserializing mapping: {}", mapping_name);
-        
-        match serde_json::from_slice(data) {
-            Ok(result) => {
-                info!("✅ Successfully deserialized mapping: {}", mapping_name);
-                Ok(result)
-            }
-            Err(e) => {
-                let error_msg = format!("Failed to deserialize {}: {}", mapping_name, e);
-                error!("❌ {}", error_msg);
-                info!("🔄 Using default value for {} due to deserialization error", mapping_name);
-                Ok(T::default())
-            }
-        }
-    }
-
-    /// Store mapping to database
-    pub fn store_mapping<T>(
-        db_ops: &std::sync::Arc<crate::db_operations::DbOperations>,
-        mapping: &RwLock<T>,
-        key: &str,
-        mapping_name: &str,
-    ) -> Result<(), SchemaError>
-    where
-        T: serde::Serialize,
-    {
-        info!("💾 Storing mapping: {} to key: {}", mapping_name, key);
-        
-        let json = Self::serialize_mapping(mapping, mapping_name)?;
-        db_ops.store_transform_mapping(key, &json)?;
-        
-        info!("✅ Successfully stored mapping: {} to database", mapping_name);
-        Ok(())
-    }
-
     // ========== EVENT PUBLISHING UTILITIES ==========
     
     /// Publish a TransformExecuted event with consistent error handling
@@ -656,52 +342,3 @@ pub type SerializationHelper = TransformUtils;
 pub type LockHelper = TransformUtils;
 pub type DefaultValueHelper = TransformUtils;
 pub type ValidationHelper = TransformUtils;
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use serde_json::json;
-
-    #[test]
-    fn test_json_to_value() {
-        let json_val = json!({"key": "value"});
-        let result = TransformUtils::json_to_value(json_val.clone());
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), json_val);
-    }
-
-    #[test]
-    fn test_validate_and_convert_string() {
-        let json_val = JsonValue::String("test".to_string());
-        let result = TransformUtils::validate_and_convert(json_val.clone(), "string", "test_field");
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), json_val);
-    }
-
-    #[test]
-    fn test_string_to_json_value_with_inference() {
-        assert_eq!(TransformUtils::string_to_json_value("true", true), JsonValue::Bool(true));
-        assert_eq!(TransformUtils::string_to_json_value("42", true), JsonValue::Number(serde_json::Number::from(42)));
-        assert_eq!(TransformUtils::string_to_json_value("hello", true), JsonValue::String("hello".to_string()));
-    }
-
-    #[test]
-    fn test_default_value_helper() {
-        assert_eq!(TransformUtils::get_default_value_for_field("input1"), JsonValue::Number(serde_json::Number::from(42)));
-        assert_eq!(TransformUtils::get_default_value_for_field("active"), JsonValue::Bool(true));
-    }
-
-    #[test]
-    fn test_infer_type_from_field_name() {
-        assert_eq!(TransformUtils::infer_type_from_field_name("user_id"), "string");
-        assert_eq!(TransformUtils::infer_type_from_field_name("age_count"), "integer");
-        assert_eq!(TransformUtils::infer_type_from_field_name("weight_value"), "float");
-        assert_eq!(TransformUtils::infer_type_from_field_name("is_active"), "boolean");
-    }
-
-    #[test]
-    fn test_validate_field_name() {
-        assert!(TransformUtils::validate_field_name("Schema.field", "test").is_ok());
-        assert!(TransformUtils::validate_field_name("invalid_field_name", "test").is_err());
-    }
-}
